chore(gcp): use fetched results for sql backups, update parent relationships
