+++ scrapers/gcp/gcp.go
@@ -10,7 +10,6 @@ import (
 	asset "cloud.google.com/go/asset/apiv1"
 	"cloud.google.com/go/asset/apiv1/assetpb"
 	"github.com/Jeffail/gabs/v2"
-	"github.com/flanksource/duty/models"
 	"github.com/flanksource/duty/types"
 	uuidV5 "github.com/gofrs/uuid/v5"
 	"github.com/google/uuid"
@@ -82,18 +81,6 @@ func getRegionFromZone(zone string) string {
 		return ""
 	}
 	return strings.Join(parts[:2], "-")
-}
-
-func getFieldValue(data *structpb.Struct, keys []string) string {
-	for _, field := range keys {
-		if value, ok := data.Fields[field]; ok {
-			if strValue := value.GetStringValue(); strValue != "" {
-				return strValue
-			}
-		}
-	}
-
-	return ""
 }
 
 func parseResourceData(data *structpb.Struct) ResourceData {
@@ -175,21 +162,6 @@ var defaultIgnoreList = []string{
 func generateConsistentID(input string) uuid.UUID {
 	gen := uuidV5.NewV5(uuidV5.NamespaceOID, input)
 	return uuid.UUID(gen)
-}
-
-func parseGCPMember(member string) (userType, name, email string, found bool) {
-	if after, ok := strings.CutPrefix(member, "user:"); ok {
-		email = after
-		return "User", email, email, true
-	} else if after, ok := strings.CutPrefix(member, "serviceAccount:"); ok {
-		email = after
-		return "ServiceAccount", email, email, true
-	} else if after, ok := strings.CutPrefix(member, "group:"); ok {
-		name = after
-		return "Group", name, name, true
-	}
-
-	return "", member, "", false
 }
 
 var unwantedFields = []string{
@@ -299,6 +271,11 @@ func (gcp Scraper) FetchAllAssets(ctx *GCPContext, config v1.GCP) (v1.ScrapeResu
 		}
 
 		relationships := RelationshipResolver(configType, rd)
+		// Add project as parent (if multiple parents are present, we use first available)
+		relationships.Parents = append(relationships.Parents, v1.ConfigExternalKey{
+			Type:       v1.GCPProject,
+			ExternalID: config.Project,
+		})
 
 		res := v1.ScrapeResult{
 			BaseScraper:         config.BaseScraper,
@@ -313,7 +290,9 @@ func (gcp Scraper) FetchAllAssets(ctx *GCPContext, config v1.GCP) (v1.ScrapeResu
 			Labels:              rd.Labels,
 			Tags:                tags,
 			Properties:          []*types.Property{getLink(rd)},
-			RelationshipResults: relationships,
+			RelationshipResults: relationships.Relationships,
+			Children:            relationships.Children,
+			Parents:             relationships.Parents,
 		}
 
 		if rd.ID != "" {
@@ -322,119 +301,6 @@ func (gcp Scraper) FetchAllAssets(ctx *GCPContext, config v1.GCP) (v1.ScrapeResu
 
 		results = append(results, res)
 	}
-
-	return results, nil
-}
-
-// FetchIAMPolicies scrapes external users and roles.
-func (gcp Scraper) FetchIAMPolicies(ctx *GCPContext, config v1.GCP) (v1.ScrapeResults, error) {
-	var results v1.ScrapeResults
-
-	req := &assetpb.ListAssetsRequest{
-		Parent:      fmt.Sprintf("projects/%s", config.Project),
-		ContentType: assetpb.ContentType_IAM_POLICY,
-	}
-
-	assetClient, err := asset.NewClient(ctx, ctx.ClientOpts...)
-	if err != nil {
-		return nil, fmt.Errorf("error creating asset client for IAM policies: %w", err)
-	}
-	defer func() {
-		if err := assetClient.Close(); err != nil {
-			ctx.Warnf("gcp iam policies: failed to close asset client: %v", err)
-		}
-	}()
-
-	// Track unique roles and users to avoid duplicates
-	uniqueRoles := make(map[uuid.UUID]models.ExternalRole)
-	uniqueUsers := make(map[uuid.UUID]models.ExternalUser)
-	var configAccesses []v1.ExternalConfigAccess
-
-	it := assetClient.ListAssets(ctx, req)
-	for {
-		asset, err := it.Next()
-		if err == iterator.Done {
-			break
-		} else if err != nil {
-			return nil, fmt.Errorf("error listing IAM policies: %w", err)
-		}
-
-		if asset.IamPolicy == nil {
-			continue
-		}
-
-		// Extract resource ID from asset name (e.g., "//compute.googleapis.com/projects/project-id/zones/us-central1-a/instances/instance-name")
-		resourceID := asset.Name
-		if resourceID == "" {
-			continue
-		}
-
-		for _, binding := range asset.IamPolicy.Bindings {
-			// bind.Role could be
-			// global: roles/cloudasset.owner
-			// custom: projects/aditya-461913/roles/mycustomroleaditya (project scoped)
-			roleID := generateConsistentID(binding.Role)
-			if _, exists := uniqueRoles[roleID]; !exists {
-				role := models.ExternalRole{
-					ID:        roleID,
-					Name:      binding.Role,
-					AccountID: config.Project,
-					ScraperID: ctx.ScrapeConfig().GetPersistedID(),
-					RoleType:  lo.Ternary(strings.HasPrefix(binding.Role, "roles/"), "Global", "Custom"),
-				}
-
-				if strings.HasPrefix(binding.Role, "roles/") {
-					role.RoleType = "Global"
-				} else {
-					role.RoleType = "Custom"
-
-					// FIXME: Only custom roles should be tied to an account and scraper
-				}
-
-				uniqueRoles[roleID] = role
-			}
-
-			for _, member := range binding.Members {
-				userType, name, email, found := parseGCPMember(member)
-				if !found {
-					continue
-				}
-
-				userID := generateConsistentID(email)
-				if _, exists := uniqueUsers[userID]; !exists {
-					externalUser := models.ExternalUser{
-						ID:        userID,
-						Name:      name,
-						ScraperID: lo.FromPtr(ctx.ScrapeConfig().GetPersistedID()),
-						AccountID: config.Project,
-						CreatedAt: time.Now(), // We don't have this information
-						UserType:  userType,
-					}
-					if email != "" {
-						externalUser.Email = &email
-					}
-					uniqueUsers[userID] = externalUser
-				}
-			}
-		}
-	}
-
-	var externalRoles []models.ExternalRole
-	for _, role := range uniqueRoles {
-		externalRoles = append(externalRoles, role)
-	}
-
-	var externalUsers []models.ExternalUser
-	for _, user := range uniqueUsers {
-		externalUsers = append(externalUsers, user)
-	}
-
-	results = append(results, v1.ScrapeResult{
-		BaseScraper:   config.BaseScraper,
-		ExternalRoles: externalRoles,
-		ExternalUsers: externalUsers,
-		ConfigAccess:  configAccesses,
-	})
 
 	return results, nil
 }
@@ -500,7 +366,13 @@ func (gcp Scraper) Scrape(ctx api.ScrapeContext) v1.ScrapeResults {
 	return processResults(allResults)
 }
 
-func RelationshipResolver(assetType string, rd ResourceData) []v1.RelationshipResult {
+type relationshipResults struct {
+	Parents       []v1.ConfigExternalKey
+	Children      []v1.ConfigExternalKey
+	Relationships []v1.RelationshipResult
+}
+
+func RelationshipResolver(assetType string, rd ResourceData) relationshipResults {
 	switch assetType {
 	case v1.GCPInstance:
 		return resolveGCPInstanceRelationships(rd)
@@ -509,26 +381,26 @@ func RelationshipResolver(assetType string, rd ResourceData) []v1.RelationshipRe
 	case v1.GCPGKECluster:
 		return resolveGCPGKEClusterRelationships(rd)
 	}
-	return nil
+	return relationshipResults{}
 }
 
-func resolveGCPInstanceRelationships(rd ResourceData) (r []v1.RelationshipResult) {
+func resolveGCPInstanceRelationships(rd ResourceData) (r relationshipResults) {
 	data := rd.Raw
 	b, _ := data.MarshalJSON()
 	p, _ := gabs.ParseJSON(b)
 	selfExternalID := v1.ExternalID{ExternalID: data.Fields["selfLink"].GetStringValue(), ConfigType: v1.GCPInstance}
 	for _, ni := range p.Search("networkInterfaces").Children() {
 		subnet := fmt.Sprint(ni.Path("subnetwork").Data())
-		r = append(r, v1.RelationshipResult{
-			ConfigExternalID:  v1.ExternalID{ExternalID: subnet, ConfigType: v1.GCPSubnet, ScraperID: "all"},
-			RelatedExternalID: selfExternalID,
-			Relationship:      "InstanceSubnet",
+		r.Parents = append(r.Parents, v1.ConfigExternalKey{
+			ExternalID: subnet,
+			Type:       v1.GCPSubnet,
+			ScraperID:  "all",
 		})
 	}
 
 	for _, disk := range p.Search("disks").Children() {
 		diskLink := fmt.Sprint(disk.Path("source").Data())
-		r = append(r, v1.RelationshipResult{
+		r.Relationships = append(r.Relationships, v1.RelationshipResult{
 			ConfigExternalID:  selfExternalID,
 			RelatedExternalID: v1.ExternalID{ExternalID: diskLink, ConfigType: v1.GCPDisk},
 			Relationship:      "InstanceDisk",
@@ -536,32 +408,39 @@ func resolveGCPInstanceRelationships(rd ResourceData) (r []v1.RelationshipResult
 	}
 
 	if cluster, exists := rd.Labels["goog-k8s-cluster-name"]; exists {
-		r = append(r, v1.RelationshipResult{
+		r.Relationships = append(r.Relationships, v1.RelationshipResult{
 			ConfigExternalID:  v1.ExternalID{ExternalID: cluster, ConfigType: v1.GCPGKECluster},
 			RelatedExternalID: selfExternalID,
 			Relationship:      "GKEInstance",
 		})
+
+		r.Relationships = append(r.Relationships, v1.RelationshipResult{
+			ConfigExternalID:  selfExternalID,
+			RelatedExternalID: v1.ExternalID{ExternalID: "Kubernetes/Node//" + rd.Name, ConfigType: "Kubernetes::Node", ScraperID: "all"},
+			Relationship:      "InstanceKuberenetesNode",
+		})
+
 	}
 	return r
 }
 
-func resolveGCPSubnetRelationships(rd ResourceData) (r []v1.RelationshipResult) {
-	selfExternalID := v1.ExternalID{ExternalID: rd.URL, ConfigType: v1.GCPSubnet}
+func resolveGCPSubnetRelationships(rd ResourceData) (r relationshipResults) {
 	if network := rd.Raw.Fields["network"].GetStringValue(); network != "" {
-		r = append(r, v1.RelationshipResult{
-			ConfigExternalID:  v1.ExternalID{ExternalID: network, ConfigType: v1.GCPNetwork},
-			RelatedExternalID: selfExternalID,
+		r.Parents = append(r.Parents, v1.ConfigExternalKey{
+			ExternalID: network,
+			Type:       v1.GCPNetwork,
+			ScraperID:  "all",
 		})
 	}
 	return r
 }
 
-func resolveGCPGKEClusterRelationships(rd ResourceData) (r []v1.RelationshipResult) {
-	selfExternalID := v1.ExternalID{ExternalID: rd.URL, ConfigType: v1.GCPGKECluster}
+func resolveGCPGKEClusterRelationships(rd ResourceData) (r relationshipResults) {
 	if network := rd.Raw.Fields["network"].GetStringValue(); network != "" {
-		r = append(r, v1.RelationshipResult{
-			ConfigExternalID:  v1.ExternalID{ExternalID: network, ConfigType: v1.GCPNetwork, ScraperID: "all"},
-			RelatedExternalID: selfExternalID,
+		r.Parents = append(r.Parents, v1.ConfigExternalKey{
+			ExternalID: network,
+			Type:       v1.GCPNetwork,
+			ScraperID:  "all",
 		})
 	}
 	return r
